requirements.txt
absl-py==2.0.0
accelerate==0.24.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.3
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
distro==1.9.0
dnspython==2.4.2
env==0.1.0
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flask==3.0.0
flatbuffers==23.5.26
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.19.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.29.0
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter-client==8.6.0
jupyter-core==5.7.1
keras==2.14.0
langcodes==3.3.0
libclang==16.0.6
loguru==0.7.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
openai==1.12.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.0.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.1
pydantic-core==2.14.3
pygments==2.17.2
pymongo==4.6.0
pypdf2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.2.1
tiktoken==0.5.2
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.35.2
typer==0.9.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.7
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
werkzeug==3.0.1
win32-setctime==1.1.0
wrapt==1.14.1
zipp==3.17.0
